refactor: move retryWithBackoff to local KV sample utility
retryWithBackoff is now a local helper (utils.ts) within each KV package's
test/public/samples/ directory. As a localHelper import (relative path), the
compiler includes it in the published sample output — no forPublishing wrapper
needed. This keeps the retry logic visible and educational for customers.
