+++ sdk/keyvault/keyvault-secrets/test/public/samples/backupAndRestore.spec.ts
@@ -10,7 +10,8 @@ import { SecretClient } from "../../../src/index.js";
 import { DefaultAzureCredential } from "@azure/identity";
 import { createTestCredential } from "@azure-tools/test-credential";
 import { Recorder, assertEnvironmentVariable } from "@azure-tools/test-recorder";
-import { forPublishing, retryWithBackoff } from "@azure-tools/test-publishing";
+import { forPublishing } from "@azure-tools/test-publishing";
+import { retryWithBackoff } from "./utils.js";
 import { describe, it, beforeEach, afterEach } from "vitest";
 // Load the .env file if it exists
 import "dotenv/config";
@@ -80,10 +81,7 @@ describe("backupAndRestore", () => {
     const backupContents = await readFile("secret_backup.dat");
     // @ts-preserve-whitespace
     // Restore the secret
-    const result = await forPublishing(
-      retryWithBackoff(() => client.restoreSecretBackup(backupContents)),
-      () => client.restoreSecretBackup(backupContents),
-    );
+    const result = await retryWithBackoff(() => client.restoreSecretBackup(backupContents));
     console.log("Restored secret: ", result);
     // @snippet-end ReadmeSampleRestoreSecret
