+++ sdk/test-utils/test-publishing/src/index.ts
@@ -15,5 +15,3 @@
 
 export { forPublishing } from "./forPublishing.js";
 export { sampleOnly } from "./sampleOnly.js";
-export { retryWithBackoff } from "./retryWithBackoff.js";
-export type { RetryWithBackoffOptions } from "./retryWithBackoff.js";
